Hide `lookupConstructor` in scala-2 src
