+++ src/main/scala/io/findify/flinkadt/api/serializer/ScalaCaseClassSerializer.scala
@@ -22,9 +22,6 @@ import org.apache.flink.api.common.typeutils.{TypeSerializer, TypeSerializerSnap
 
 import java.io.ObjectInputStream
 
-import scala.annotation.nowarn
-import scala.reflect.runtime.universe
-
 /** This is a non macro-generated, concrete Scala case class serializer.
   * Copied from Flink 1.14 without `SelfResolvingTypeSerializer`.
   */
@@ -52,37 +49,6 @@ class ScalaCaseClassSerializer[T <: Product](
   }
 }
 
-object ScalaCaseClassSerializer {
-  @nowarn("msg=(eliminated by erasure)|(explicit array)")
-  def lookupConstructor[T](cls: Class[T]): Array[AnyRef] => T = {
-    val rootMirror  = universe.runtimeMirror(cls.getClassLoader)
-    val classSymbol = rootMirror.classSymbol(cls)
-
-    require(
-      classSymbol.isStatic,
-      s"""
-         |The class ${cls.getSimpleName} is an instance class, meaning it is not a member of a
-         |toplevel object, or of an object contained in a toplevel object,
-         |therefore it requires an outer instance to be instantiated, but we don't have a
-         |reference to the outer instance. Please consider changing the outer class to an object.
-         |""".stripMargin
-    )
-
-    val primaryConstructorSymbol = classSymbol.toType
-      .decl(universe.termNames.CONSTRUCTOR)
-      .alternatives
-      .collectFirst {
-        case constructorSymbol: universe.MethodSymbol if constructorSymbol.isPrimaryConstructor =>
-          constructorSymbol
-      }
-      .head
-      .asMethod
-
-    val classMirror             = rootMirror.reflectClass(classSymbol)
-    val constructorMethodMirror = classMirror.reflectConstructor(primaryConstructorSymbol)
-
-    (arr: Array[AnyRef]) => {
-      constructorMethodMirror.apply(arr: _*).asInstanceOf[T]
-    }
-  }
-}
+// Indirectly enable constructor look-up.
+// Underlying implementation is major version-specific (Scala 2 vs. Scala 3).
+object ScalaCaseClassSerializer extends ConstructorCompat
